+++ tests/Justimmo/Tests/Wrapper/V1/RealtyWrapperTest.php
@@ -46,6 +46,8 @@ class RealtyWrapperTest extends TestCase
         $this->assertInstanceOf('\DateTime', $entry->getUpdatedAt(null));
         $this->assertEquals('2015-12-17 11:08:39', $entry->getUpdatedAt());
         $this->assertEquals('17.12.2015', $entry->getUpdatedAt('d.m.Y'));
+        $this->assertEquals(16.4100297, $entry->getLongitudePrecise());
+        $this->assertEquals(48.2545373, $entry->getLatitudePrecise());
 
         $entry = $list[1];
         $this->assertInstanceOf('\Justimmo\Model\Realty', $entry);
@@ -117,6 +119,7 @@ class RealtyWrapperTest extends TestCase
         $this->assertEquals($objekt->getParkingArea(), 36.85);
 
         $this->assertEquals(450000, $objekt->getPurchasePrice());
+        $this->assertEquals(3000, $objekt->getPurchasePricePerSqm());
         $this->assertEquals(500, $objekt->getAdditionalCharges());
         $this->assertEquals(126, $objekt->getHeatingCosts());
         $this->assertEquals('EUR', $objekt->getCurrency());
@@ -264,6 +267,7 @@ class RealtyWrapperTest extends TestCase
 
         $this->assertEquals(1, $objekt->getStyleOfBuildingId());
 
-
+        $this->assertEquals(16.4100297, $objekt->getLongitudePrecise());
+        $this->assertEquals(48.2545373, $objekt->getLatitudePrecise());
     }
 }